Fixed issues and splitted background file to make it cleaner
