+++ public/js/background.js
@@ -1,5 +1,5 @@
+import { pomodoroDuration, breakDuration, timerRunning, startTimer, resetTimer, pauseTimer, sendTimer, logAchievement } from './pomodoroTimer.js';
 
-let extensionEnabled = true;
 
 // Event listener for when the extension is installed
 chrome.runtime.onInstalled.addListener(() => {
@@ -20,27 +20,13 @@ chrome.tabs.onUpdated.addListener((tabId, changeInfo, tab) => {
   });
 });
 
-// Pomodoro timer variables
-let countdown;
-let timerDuration;
-let timerRunning = false;
-let onBreak = false;
-let pomodoroDuration = 25 * 60; // default Pomodoro duration
-let breakDuration = 5 * 60; // default break duration
-let isPaused = false;
-let pausedTime = 0;
-let startDate;
-let endTime;
-let startTime;
-let totalTimeElapsed;
 
-// Event listener for messages from the popup
+
+// Event listener for messages from the pomodoro timer
 chrome.runtime.onMessage.addListener((msg, sender, sendResponse) => {
   switch (msg.command) {
     case 'start':
-      pomodoroDuration = msg.pomodoroDuration * 60 || pomodoroDuration;
-      breakDuration = msg.breakDuration * 60 || breakDuration;
-      startTimer(pomodoroDuration);
+      startTimer(msg.pomodoroDuration * 60, msg.breakDuration * 60);
       break;
     case 'reset':
       resetTimer();
@@ -62,138 +48,4 @@ chrome.runtime.onMessage.addListener((msg, sender, sendResponse) => {
       console.error('Unrecognized command');
   }
 });
-
-// Function to start the timer
-function startTimer(duration) {
-  if (!timerRunning) {
-    timerDuration = isPaused ? pausedTime : duration; // Start from paused time if timer was paused
-    timerRunning = true;
-    isPaused = false;
-    chrome.alarms.create('pomodoroTimer', { delayInMinutes: 1 / 60 });
-    updateTimer();
-  }
-}
-
-// Function to update the timer every second
-function updateTimer() {
-  countdown = setInterval(() => {
-    timerDuration--;
-    if (timerDuration <= 0) {
-      clearInterval(countdown);
-      timerRunning = false;
-      onBreak = !onBreak;
-      displayNotification();
-      if (!onBreak) {
-        resetTimer(); // Reset the timer when the break time is finished
-        
-      }
-    }
-    sendTimer();
-  }, 1000);
-}
-
-// Function to reset the timer
-function resetTimer() {
-  clearInterval(countdown);
-  timerRunning = false;
-  onBreak = false;
-  isPaused = false;
-  timerDuration = pomodoroDuration;
-  sendTimer();
-}
-
-// Function to pause the timer
-function pauseTimer() {
-  if (timerRunning) {
-    clearInterval(countdown);
-    isPaused = true;
-    pausedTime = timerDuration;
-    sendTimer();
-  }
-}
-
-//function to open achievement input page
-function openInputPage() {
-    chrome.windows.create({ url: 'input.html', type: 'popup', width: 500, height: 600 });
-    startDate = new Date().toLocaleDateString('en-US', { day: 'numeric', month: 'long', year: 'numeric' });;
-    endTime = new Date().toLocaleTimeString('en-US', { hour: '2-digit', minute: '2-digit' });
-    startTime = new Date(new Date().getTime() - (pomodoroDuration * 1000)).toLocaleTimeString('en-US', { hour: '2-digit', minute: '2-digit' });
-    totalTimeElapsed = new Date(pomodoroDuration * 1000).toISOString().slice(11, 19);
-}
-
-// Function to log the achievement
-function logAchievement(achievement) {
-  
-  // Save the data to Chrome storage
-  chrome.storage.sync.get('focusSessionData', (result) => {
-    let focusSessionData = Array.isArray(result.focusSessionData) ? result.focusSessionData : [];
-    
-    const data = {
-      startDate: startDate,
-      startTime: startTime,
-      endTime: endTime,
-      totalTimeElapsed: totalTimeElapsed,
-      achievement: achievement
-    };
-    focusSessionData.push(data);
-    chrome.storage.sync.set({ focusSessionData: focusSessionData }, () => {
-      if (chrome.runtime.lastError) {
-        // Handle error
-        console.log(chrome.runtime.lastError.message);
-      } else {
-        console.log('Achievement data saved to Chrome storage.');
-      }
-    });
-  })
-}
-
-
-// Function to send the timer to the popup
-function sendTimer() {
-  const minutes = Math.floor(timerDuration / 60);
-  const seconds = timerDuration % 60;
-  chrome.runtime.sendMessage({ minutes: minutes, seconds: seconds, onBreak: onBreak }, function(response) {
-    if (chrome.runtime.lastError) {
-      // Handle error
-      console.log(chrome.runtime.lastError.message);
-    }
-  });
-}
-
-// Function to display a notification when the timer is finished
-function displayNotification() {
-  const buttonTitle = onBreak ? 'Start Break' : 'Finish session';
-  chrome.notifications.create('pomodoroNotification', {
-    type: 'basic',
-    iconUrl: '../images/zenlify_logo.png',
-    title: onBreak ? 'Take a break!' : 'The pomodoro session is over!',
-    message: onBreak ? 'Take a break!' : 'The pomodoro session is over!',
-    buttons: [
-      { title: buttonTitle }
-    ],
-    requireInteraction: true, // Prevent the notification from disappearing until the user clicks on the button
-    priority: 2
-  }, (notificationId) => {
-    // Event listener for when the notification button is clicked
-    chrome.notifications.onButtonClicked.addListener((clickedNotificationId, buttonIndex) => {
-      if (clickedNotificationId === notificationId && buttonIndex === 0) {
-        if (!onBreak) {
-          // Handle "Finish Timer" button click
-          resetTimer();
-          
-        } else {
-          // Handle "Start Break" button click
-          openInputPage();
-          startBreakTimer();
-        }
-      }
-    });
-  });
-}
-
-function startBreakTimer() {
-  timerDuration = breakDuration;
-  startTimer(timerDuration);
-}
-
