+++ public/manifest.json
@@ -5,7 +5,8 @@
   "description": "Block distracting content on the web",
   "options_page": "index.html",
   "background": {
-    "service_worker": "js/background.js"
+    "service_worker": "js/background.js",
+    "type": "module"
   },
   "action": {
     "default_popup": "popup.html",
